Simplified finding 'StartingRoom' position handling
Replaced _game_map dictionary with pos tuple to store 'StartingRoom' position. 
Replaced associated code which handles assigning values to _game_map with if statement to store position of 'StartingRoom' into pos when the tile_name is encountered during parsing of the map.
Removed redundant function that handles assigning of position.